requirements.txt
feedparser
requests
psycopg2-binary
boto3
PyYAML
prefect
python-dotenv